perf: concurrency and memory improvements for execution layer
Description

This PR refers to the code optimization done by `BSC`, which mainly includes the following parts:
1. Do BlockBody verification concurrently.
2. Do the calculation of intermediate root concurrently.
3. Commit the MPTs concurrently.
4. Preload accounts before processing blocks.
5. Make the snapshot layers configurable.
6. Reuse some objects to reduce GC.
7. Add shared_pool for `stateDB` to improve cache usage.

References
- https://github.com/bnb-chain/bsc/pull/257
- https://github.com/bnb-chain/bsc/pull/792

---------

+++ core/state/journal.go
@@ -43,7 +43,8 @@ type journal struct {
 // newJournal creates a new initialized journal.
 func newJournal() *journal {
 	return &journal{
-		dirties: make(map[common.Address]int),
+		dirties: make(map[common.Address]int, defaultNumOfSlots),
+		entries: make([]journalEntry, 0, defaultNumOfSlots),
 	}
 }
 
@@ -266,7 +267,9 @@ func (ch accessListAddAccountChange) revert(s *StateDB) {
 		(addr) at this point, since no storage adds can remain when come upon
 		a single (addr) change.
 	*/
-	s.accessList.DeleteAddress(*ch.address)
+	if s.accessList != nil {
+		s.accessList.DeleteAddress(*ch.address)
+	}
 }
 
 func (ch accessListAddAccountChange) dirtied() *common.Address {
@@ -274,7 +277,9 @@ func (ch accessListAddAccountChange) dirtied() *common.Address {
 }
 
 func (ch accessListAddSlotChange) revert(s *StateDB) {
-	s.accessList.DeleteSlot(*ch.address, *ch.slot)
+	if s.accessList != nil {
+		s.accessList.DeleteSlot(*ch.address, *ch.slot)
+	}
 }
 
 func (ch accessListAddSlotChange) dirtied() *common.Address {
